0x02. Python - Async Comprehension Mandatory Task 0

+++ 0x02-python_async_comprehension/0-async_generator.py
@@ -7,11 +7,11 @@ import asyncio
 from random import randint
 
 
-async def async_generator() -> AsyncGenerator[int, None]:
+async def async_generator() -> AsyncGenerator[float, None]:
     """
     Returns a random number between 0 and 10
     """
 
     for _ in range(10):
         await asyncio.sleep(1)
-        yield randint(0, 10)
+        yield float(randint(0, 10))
